requirements.txt
django
djangorestframework
django-rest-knox
pika
psycopg2